config_modelsummary gets startup_message  argument

+++ R/config_modelsummary.R
@@ -11,29 +11,28 @@ config_modelsummary <- function(
     factory_latex,
     factory_html,
     factory_markdown,
+    startup_message,
     reset = FALSE) {
 
+    # missing() is useful so we can call config_modelsummary() with just one
+    # argument without changin all the other ones
+
     dn <- tools::R_user_dir(package = "modelsummary", which = "config")
-    if (!dir.exists(dn)) dir.create(dn, recursive = TRUE)
-
     fn <- file.path(dn, "config.rds")
-    if (!file.exists(fn)) {
-        config <- list()
-    } else {
-        config <- readRDS(fn)
-    }
 
-    if (isTRUE(reset)) {
-        hush(unlink(fn))
-        insight::format_warning("`modelsummary` returned to default settings.")
-    }
+    if (isTRUE(reset) && file.exists(fn)) hush(unlink(fn))
 
-    if (missing(factory_default) && missing(factory_latex) && missing(factory_html) && missing(factory_markdown)) {
+    config <- config_read()
+
+    if (missing(factory_default) && missing(factory_latex) && missing(factory_html) && missing(factory_markdown) && missing(startup_message)) {
         return(config)
     }
 
     if (!missing(factory_default)) {
-        checkmate::assert_choice(factory_default, c("tinytable", "kableExtra", "gt", "huxtable", "flextable", "DT", "markdown"), null.ok = TRUE)
+        checkmate::assert_choice(
+            factory_default, 
+            c("tinytable", "kableExtra", "gt", "huxtable", "flextable", "DT", "markdown"),
+            null.ok = TRUE)
         if (isTRUE(factory_default == "tinytable")) insight::check_if_installed("tinytable")
         if (isTRUE(factory_default == "kableExtra")) insight::check_if_installed("kableExtra")
         if (isTRUE(factory_default == "gt")) insight::check_if_installed("gt")
@@ -46,11 +45,18 @@ config_modelsummary <- function(
     if (!missing(factory_latex)) {
         config[["factory_latex"]] <- factory_latex
     }
+
     if (!missing(factory_html)) {
         config[["factory_html"]] <- factory_html
     }
+
     if (!missing(factory_markdown)) {
         config[["factory_markdown"]] <- factory_markdown
+    }
+
+    if (!missing(startup_message)) {
+        config[["startup_message"]] <- startup_message
+        checkmate::assert_flag(startup_message)
     }
 
     saveRDS(config, file = fn)
@@ -58,7 +64,33 @@ config_modelsummary <- function(
 
 
 config_get <- function(x) {
-    fn <- file.path(tools::R_user_dir(package = "modelsummary", which = "config"), "config.rds")
-    if (!file.exists(fn)) return(NULL)
-    return(readRDS(fn)[[x]])
+    config <- config_read()
+    return(config[[x]])
+}
+
+
+config_delete <- function() {
+    dn <- tools::R_user_dir(package = "modelsummary", which = "config")
+    fn <- file.path(dn, "config.rds")
+    if (file.exists(fn)) hush(unlink(fn))
+    insight::format_warning("`modelsummary` returned to default settings.")
+}
+
+
+config_read <- function() {
+    dn <- tools::R_user_dir(package = "modelsummary", which = "config")
+    if (!dir.exists(dn)) dir.create(dn, recursive = TRUE)
+    fn <- file.path(dn, "config.rds")
+    if (!file.exists(fn)) {
+        config <- list(
+            factory_default = "tinytable",
+            factory_latex = "tinytable",
+            factory_html = "tinytable",
+            factory_markdown = "tinytable",
+            startup_message = TRUE
+        )
+    } else {
+        config <- readRDS(fn)
+    }
+    return(invisible(config))
 }
